pc: push experimental 64-bit bringup forward
Fix strict-compiler issues and a few pointer-width hotspots so the guarded macOS build reaches later JSystem blockers.

+++ include/dolphin/card/CARDCheck.h
@@ -2,6 +2,6 @@
 #define _DOLPHIN_CARDCHECK_H_
 
 s32 CARDCheckAsync(s32 chan, CARDCallback callback);
-long CARDCheck(long chan);
+s32 CARDCheck(s32 chan);
 
 #endif // _DOLPHIN_CARDCHECK_H_
