+++ include/dolphin/card/CARDDelete.h
@@ -2,7 +2,7 @@
 #define _DOLPHIN_CARDDELETE_H_
 
 s32 CARDFastDeleteAsync(s32 chan, s32 fileNo, CARDCallback callback);
-long CARDFastDelete(long chan, long fileNo);
+s32 CARDFastDelete(s32 chan, s32 fileNo);
 s32 CARDDeleteAsync(s32 chan, const char *fileName, CARDCallback callback);
 s32 CARDDelete(s32 chan, const char *fileName);
